Fix for terminology: does not fallback on ascii when no wallpaper is found
When no image is set and there is no background image (or a `.xml` is detected) and the image option is set to `wall` (default), neoneofetch does not fallback on ascii mode.
The issue comes from the line 2310:

    if [ ! -f "$img" ] || [ ${#term_size} -le 5 ] && [ "$image_backend" != "tycat" ]; then

as `||` and `&&` have the same priority, bash does not evaluate `[ ${#term_size} -le 5 ] && [ "$image_backend" != "tycat" ]` first.

Solution is simply this

    if [ ! -f "$img" ] || ([ ${#term_size} -le 5 ] && [ "$image_backend" != "tycat" ]); then