Fix --ascii_distro ubuntu-studio
The ascii/distro directory has an ubuntu-studio file, which is
documented in the --help text, but when I try to use that option, it
falls back to the default Ubuntu logo.

This change adds a case for Ubuntu-Studio to set ascii_file and the
colors based on the logo at https://ubuntustudio.org.