Show GPU subsystem name if possible. Fixes #1490 (#1648)
After this change, if a subsystem name exists for a GPU, we display that
name instead of the device name. This should fix the issue mentioned
in #1490.

For example, a card that would be shown as:
    GPU: AMD ATI Radeon RX 470/480/570/570X/580/580X/590
before will now be shown as:
    GPU: AMD ATI Radeon RX 480 4GB
This fallback logic is implemented in the `awk` command.

Additionally, the PCI address is now stripped from the awk output
because it never is desirable to print it. It is, however, still
kept internally in `awk` to distinguish multiple cards of the same model.